perf(gas): optimize stage per wallet limits gas
- pack into aux on the erc721a ownership storage
- caveat: limits # of stages to 4 (or 8 if max limit per wallet of 255
  is acceptable)

+++ contracts/IERC721M.sol
@@ -79,7 +79,7 @@ interface IERC721M is IERC721AQueryable {
         view
         returns (
             MintStageInfo memory,
-            uint32,
+            uint16,
             uint256
         );
